fix: filter district

+++ tables/community_farm.sql
@@ -77,7 +77,7 @@ CREATE TABLE community_crop_reports(
     updatedAt timestamp DEFAULT CURRENT_TIMESTAMP,
     is_archived BOOLEAN DEFAULT false,
     is_first_report BOOLEAN default TRUE,
-    kilogram INT default 0,
+    kilogram NUMERIC default 0,
     harvested_by INT REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (crop_id) REFERENCES community_farms_crops(id) ON DELETE CASCADE, 
     FOREIGN KEY (farmid) REFERENCES community_farms(id) ON DELETE CASCADE,
